+++ sw.js
@@ -89,7 +89,8 @@
 // 2026-06-11 v92: v7.427 버그 리포터 JS 외부화.
 // 2026-06-11 v93: v7.428 버그 리포터 CSS 외부화.
 // 2026-06-11 v94: v7.429 함선/광물 도감 모달 CSS/JS 외부화.
-const CACHE_NAME = 'mars-v94';
+// 2026-06-11 v95: v7.430 공통 게임 다이얼로그 JS 외부화.
+const CACHE_NAME = 'mars-v95';
 const STATIC_ASSETS = [
   '/',
   '/manifest.json'